so changed it a bit so you just get a cute list of pokemans...
HOWEVER

yeah for longer lists made like... the code I'm using to delete is barely functional and often not... I think the lesson is don't use a recycler view for things that are going to change in a super dynamic way... or maybe the lesson is handle deletions carefully and slowly.

there's a lesson probably. I need an intermediate course in recycleviews.